simplify has()
Not guarding against functions and aliases anymore is intentional, it
means we can do things like define a function to wrap around an external
command and still have it be seen by has(). In hyfetch I added a wrapper
function around sed to replace any usage of -E with -r if we were using
GNU sed, which fixes issues with extremely old versions of GNU sed that
don't support -E, that kind of thing would now be seen by has(). This
shouldn't cause any issues since pfetch is run in a subshell that
doesn't have access to normal user-defined functions and aliases, but if
it does I'll revert this.